requirements.txt
# =============================================================================
# Portfolio Optimizer - Python Dependencies
# =============================================================================

# Web Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-CORS>=4.0.0
Werkzeug>=3.0.1

# Database
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data & Finance
yfinance>=0.2.33
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0

# AI/LLM Integration
openai>=1.6.1

# Document Processing (PDF/CSV parsing)
PyPDF2>=3.0.1
pdfplumber>=0.10.0
openpyxl>=3.1.2

# Data Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
cachetools>=5.3.0
python-dateutil>=2.8.2
cryptography>=41.0.0

# Optimization
cvxpy>=1.4.0

# Testing
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
statsmodels>=0.14.0
